Add hero
